+++ test/gg_git_test.dart
@@ -35,33 +35,45 @@ void main() {
         // Iterate all files in lib/src/commands
         // and check if they are added to the command runner
         // and if they are added to the help message
-        final subCommands = Directory('lib/src/commands')
-            .listSync(recursive: false)
-            .where(
-              (file) => file.path.endsWith('.dart'),
-            )
-            .map(
-              (e) => basename(e.path)
-                  .replaceAll('.dart', '')
-                  .replaceAll('_', '-')
-                  .replaceAll('gg-', ''),
-            )
-            .toList();
+        final subCommandsDart =
+            Directory('lib/src/commands').listSync(recursive: false).where(
+                  (file) => file.path.endsWith('.dart'),
+                );
 
         await capturePrint(
           ggLog: messages.add,
           code: () async => await runner.run(['ggGit', '--help']),
         );
 
-        for (final subCommand in subCommands) {
+        for (final dartFile in subCommandsDart) {
+          final fileName = basename(dartFile.path);
+          final subCommand = fileName
+              .replaceAll('.dart', '')
+              .replaceAll('_', '-')
+              .replaceAll('gg-', '');
+
+          // Check if the command has the right name
+          final fileContent = await File(dartFile.path).readAsString();
+          final hasRightName = fileContent.contains('name: \'$subCommand\'') ||
+              fileContent.contains('name = \'$subCommand\'');
+          expect(
+            hasRightName,
+            isTrue,
+            reason: '\nPlease open "$fileName" '
+                'and make sure the name is "$subCommand".',
+          );
+
+          // Make sure the command is listed within the main command
           final ggSubCommand = subCommand.pascalCase;
 
           expect(
             hasLog(messages, subCommand),
             isTrue,
             reason: '\nMissing subcommand "$ggSubCommand"\n'
-                'Please open  "lib/src/gg_git.dart" and add\n'
-                '"addSubcommand($ggSubCommand(ggLog: ggLog));',
+                'Please open  "lib/src/gg_git.dart"\n'
+                '"and add "addSubcommand($ggSubCommand(ggLog: ggLog));"\n'
+                'Make sure the name of the command is '
+                '"$subCommand".',
           );
         }
       });